Move indenting tests to own test class
